remove useless tables and columns (#92)
