+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
 
     <application
         android:name=".MaintainerApp"
-        android:allowBackup="true"
+        android:allowBackup="false"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_launcher"
