--- app/src/main/java/de/chrisbecker386/maintainer/data/entity/Precondition.kt
@@ -1,52 +1,0 @@
-/*
- * Created by Christopher Becker on 13/05/2023, 08:24
- * Copyright (c) 2023. All rights reserved.
- * Last modified 13/05/2023, 08:24
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package de.chrisbecker386.maintainer.data.entity
-
-import androidx.annotation.DrawableRes
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.ForeignKey
-import androidx.room.Index
-import androidx.room.PrimaryKey
-
-@Entity(
-    tableName = "preconditions",
-    foreignKeys = [
-        ForeignKey(
-            entity = Task::class,
-            parentColumns = ["task_id"],
-            childColumns = ["precondition_fk_task_id"],
-            onDelete = ForeignKey.CASCADE
-        )
-    ],
-    indices = [Index("precondition_fk_task_id")]
-)
-data class Precondition(
-    @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = "precondition_id")
-    val id: Int = 0,
-    @ColumnInfo(name = "precondition_title")
-    val title: String,
-    @ColumnInfo(name = "precondition_image")
-    @DrawableRes
-    val image: Int,
-    @ColumnInfo(name = "precondition_fk_task_id")
-    val taskId: Int
-)
