+++ app/src/main/java/de/chrisbecker386/maintainer/data/local/MaintainerDb.kt
@@ -22,7 +22,6 @@ package de.chrisbecker386.maintainer.data.local
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import de.chrisbecker386.maintainer.data.entity.Machine
-import de.chrisbecker386.maintainer.data.entity.Precondition
 import de.chrisbecker386.maintainer.data.entity.Section
 import de.chrisbecker386.maintainer.data.entity.Step
 import de.chrisbecker386.maintainer.data.entity.Task
@@ -34,10 +33,9 @@ import de.chrisbecker386.maintainer.data.entity.TaskCompletedDate
         Machine::class,
         Task::class,
         Step::class,
-        Precondition::class,
         TaskCompletedDate::class
     ],
-    version = 13,
+    version = 14,
     exportSchema = false
 )
 
@@ -46,6 +44,5 @@ abstract class MaintainerDb : RoomDatabase() {
     abstract val machineDao: MachineDao
     abstract val taskDao: TaskDao
     abstract val stepDao: StepDao
-    abstract val preconditionDao: PreconditionDao
     abstract val taskCompletedDao: TaskCompletedDao
 }
