+++ app/src/main/java/de/chrisbecker386/maintainer/data/model/dummy/Steps.kt
@@ -27,7 +27,6 @@ val devSteps = listOf(
         id = 1,
         order = 1,
         title = "Open lid",
-        imageRes = null,
         description = null,
         completedDate = null,
         taskId = 1
@@ -36,7 +35,6 @@ val devSteps = listOf(
         id = 2,
         order = 2,
         title = "Remove water tank",
-        imageRes = null,
         description = "Carefully lift the tank out of the machine. So that no water is spilled unnecessarily",
         completedDate = null,
         taskId = 1
@@ -45,7 +43,6 @@ val devSteps = listOf(
         id = 3,
         order = 3,
         title = "Clean water tank",
-        imageRes = null,
         description = "Clean the water tank properly with a brush, warm water and detergent. Then rinse again with clean water and dry the tank.",
         completedDate = null,
         taskId = 1
@@ -54,7 +51,6 @@ val devSteps = listOf(
         id = 4,
         order = 4,
         title = "Reinstall the water tank",
-        imageRes = null,
         description = null,
         completedDate = null,
         taskId = 1
@@ -63,7 +59,6 @@ val devSteps = listOf(
         id = 5,
         order = 5,
         title = "Close lit",
-        imageRes = null,
         description = null,
         completedDate = null,
         taskId = 1
@@ -73,7 +68,6 @@ val devSteps = listOf(
         id = 6,
         order = 1,
         title = "Open lit",
-        imageRes = null,
         description = null,
         completedDate = null,
         taskId = 2
@@ -82,7 +76,6 @@ val devSteps = listOf(
         id = 7,
         order = 2,
         title = "Empty Tank",
-        imageRes = null,
         description = "Carefully empty the tank",
         completedDate = null,
         taskId = 2
@@ -91,7 +84,6 @@ val devSteps = listOf(
         id = 8,
         order = 3,
         title = "add descaling agent to the tank",
-        imageRes = null,
         description = "Proportional to the size of the water tank: Dissolve the descaler in lukewarm water and then pour it into the water tank.",
         completedDate = null,
         taskId = 2
@@ -100,7 +92,6 @@ val devSteps = listOf(
         id = 9,
         order = 4,
         title = "uncoggling",
-        imageRes = null,
         description = "Place a sufficiently large container under the coffee outlet. Let the machine warm up and then run the entire tank through once. If the container is full, stop briefly and empty the container. If the tank is empty, stop the machine. ",
         completedDate = null,
         taskId = 2
@@ -109,7 +100,6 @@ val devSteps = listOf(
         id = 10,
         order = 5,
         title = "Flush out residues",
-        imageRes = null,
         description = "Fill the water tank again with ordinary water and let the tank run completely empty again. Carry out this procedure 2 times",
         completedDate = null,
         taskId = 2
@@ -118,7 +108,6 @@ val devSteps = listOf(
         id = 11,
         order = 6,
         title = "Refill tank",
-        imageRes = null,
         description = null,
         completedDate = null,
         taskId = 2
@@ -127,7 +116,6 @@ val devSteps = listOf(
         id = 12,
         order = 7,
         title = "Close lit",
-        imageRes = null,
         description = null,
         completedDate = null,
         taskId = 2
