+++ app/src/main/java/de/chrisbecker386/maintainer/data/model/dummy/Tasks.kt
@@ -28,7 +28,6 @@ val devTasks = listOf(
         title = "clean water tank",
         subtitle = null,
         imageRes = R.drawable.kettle_48px,
-        duration = 5,
         machineId = 1
     ),
     Task(
@@ -36,7 +35,6 @@ val devTasks = listOf(
         title = "unclogging",
         subtitle = null,
         imageRes = R.drawable.ic_launcher_foreground,
-        duration = 5,
         machineId = 1
     ),
     Task(
@@ -44,7 +42,6 @@ val devTasks = listOf(
         title = "carTask1",
         subtitle = null,
         imageRes = R.drawable.question_mark_48px,
-        duration = 5,
         machineId = 2
     )
